03-asynchronous-request-reply-pattern/brian/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
line-bot-sdk>1.1.0,<2.0.0
httpx>=0.24.0,<0.25.0
azure-storage-queue>=12.6.0,<13.0.0
azure-identity>=1.12.0,<1.13.0